#50: Move files to src. Make installable. Add CLI cmds.
