+++ src/raindrop_todoist_syncer.egg-info/PKG-INFO
@@ -1,0 +1,5 @@
+Metadata-Version: 2.1
+Name: raindrop-todoist-syncer
+Version: 0.1.0
+Summary: Converts favourited Raindrops into tasks in Todoist.
+Author: Chris Billows
