+++ tests/unit/test_raindrop_processor.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 import pytest
 
-from raindrop import RaindropsProcessor, Raindrop
+from raindrop_todoist_syncer.raindrop import RaindropsProcessor, Raindrop
 
 
 @pytest.fixture
